Separated log id + attendance id API fetching logic to a concern
